cb fast-path

+++ include/e4pipe/infinitypipe_struct.h
@@ -15,5 +15,5 @@ struct infinitypipe
     // callback
     infinitypipe_notify_fn fn;
     void *fn_arg;
-    int notify_pending;
+    size_t notify_pending;
 };
